backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
asyncpg==0.29.0
aiosqlite==0.20.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dateutil==2.9.0
httpx==0.27.0
pydantic==2.6.4
pydantic-settings==2.2.1
apscheduler==3.10.4